Search API with Test Cases
