+++ nt.webapi/src/Infrastructure/Nt.Infrastructure.Repository/Repositories/User/UserProfileRepository.cs
@@ -2,7 +2,10 @@
 using Nt.Domain.Entities.Exceptions;
 using Nt.Domain.Entities.User;
 using Nt.Domain.RepositoryContracts.User;
+using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 
 namespace Nt.Infrastructure.Data.Repositories.User
